Extract ride status verification to separate service
